+++ src/commands/data/index.ts
@@ -12,16 +12,14 @@ import { initDataModel } from "./initDataModel";
 
 export function Data(cmd): Promise<boolean> {
     const params = cleanArgs(cmd);
-    // TODO: get datamodel folder from project config descriptor file.
-    const dataModelFolder = projectConfig.datamodelFolder;
     // TODO: get filename from project config descriptor file.
-    const dataModelFile = params.filename ? params.filename : "datamodel.xml";
-    const filename = path.join(dataModelFolder, dataModelFile);
-    const dataModelCreator = new CopyHandler(assets.datamodel, filename);
+    const dataModelPath = params.sourcepath ? params.sourcepath : projectConfig.datamodelPath;
+    const sourcepath = path.join(dataModelPath);
+    const dataModelCreator = new CopyHandler(assets.datamodel, sourcepath);
     if (params.init) {
       return initDataModel(cmd, dataModelCreator);
       // TODO: save filename to project config descriptor file on successful creation.
     } else {
-      return generateModels(cmd, dataModelCreator, params.module);
+      return generateModels(cmd, dataModelCreator, params);
     }
 }