+++ src/commands/data/modelEntityFactory.ts
@@ -1,10 +1,10 @@
 import { ModelEntityReference } from "./modelEntity.reference";
 import { ModelEntityModule } from "./modelEntity.module";
 
-export function modelEntityFactory(entity: any, isModule: boolean) {
-    if (isModule) {
-        return new ModelEntityModule(entity);
+export function modelEntityFactory(entity: any, params: any) {
+    if (params.module) {
+        return new ModelEntityModule(entity, params);
     } else {
-        return new ModelEntityReference(entity);
+        return new ModelEntityReference(entity, params);
     }
 }