+++ src/index.ts
@@ -30,7 +30,8 @@ program
   .command("data")
   .description("manage datamodel, create new or generate typescript classes.\n")
   .option("-i, --init", "Initialize the datamodel file with example data.")
-  .option("-n, --filename", "create datamodel with custom filename. default: 'datamodel.xml'")
+  .option("-s, --sourcepath [path]", "create datamodel from custom filepath. ['model/datamodel.xml']", 'model/datamodel.xml')
+  .option("-t, --targetdir [path]", "create datamodel files into the target directory. ['model/']", 'model/')
   .option("-m, --module", "generate classes for ES6 modules")
   .action((cmd) => {
     Data(cmd);